==> 2 Correlations.R <==
if(!require(tidyverse)) install.packages("tidyverse"); library(tidyverse)
#source("0 General.R" %>% paste0("C:/Users/mar84qk/Dropbox/Arbeit/C10 - Gamer Fear/3 Diagnostic Generalization/3 Analysis/main Anx/", .))

requirePackage("lme4")
requirePackage("lmerTest")


# ratings.wide = read_rds("ratings.wide.rds" %>% paste0(path.rds, .)) #or path
# eyes.wide = read_rds("eyes.wide.rds" %>% paste0(path.rds, .)) #or path
# pupil.wide = read_rds("pupil.wide.rds" %>% paste0(path.rds, .)) #or path
# heart.wide = read_rds("heart.wide.rds" %>% paste0(path.rds, .)) #or path
# eda.wide = read_rds("eda.wide.rds" %>% paste0(path.rds, .)) #or path


data.wide = full_join(questionnaires, ratings.wide, by="subject") %>% 
  full_join(pupil.wide, by="subject") %>% 
  full_join(heart.wide, by="subject") %>% 
  full_join(eda.wide, by="subject") %>% 
  inner_join(eyes.wide, by=c("subject", "SPAI", "STAI")) #only keep subjects that have valid eye-tracking data

#all(data.wide == read_rds("data.wide.rds" %>% paste0(path.rds, .)), na.rm=T) #check equivalence of processing
#colSums(data.wide != read_rds("data.wide.rds" %>% paste0(path.rds, .)), na.rm=T) %>% bind_rows() %>% pivot_longer(everything()) %>% filter(value!=0)
#data.wide %>% write_rds("data.wide.rds" %>% paste0(path.rds, .))


# Screening Quality -------------------------------------------------------
#note: use questionnaires data.frame for as much data as possible. Use data.wide instead for final sample
questionnaires %>% transmute(recruitment = ifelse(Screening_vorher %>% is.na(), "screening", "other")) %>% 
  count(recruitment) %>% mutate(prop = n / sum(n))

questionnaires %>% with(cor.test(Screening_vorher, Screening_nachher, alternative="greater")) %>% correlation_out()
#questionnaires %>% ggplot(aes(x=Screening_vorher, y=Screening_nachher)) + geom_smooth(method="lm", color="black") + geom_point(size=2) + myGgTheme

questionnaires %>% with(cor.test(SPAI, Screening_vorher, alternative="greater")) %>% correlation_out()
#questionnaires %>% ggplot(aes(x=SPAI, y=Screening_vorher)) + geom_smooth(method="lm", color="black") + geom_point(size=2) + myGgTheme

questionnaires %>% with(cor.test(SPAI, Screening_nachher, alternative="greater")) %>% correlation_out()
#questionnaires %>% ggplot(aes(x=SPAI, y=Screening_nachher)) + geom_smooth(method="lm", color="black") + geom_point(size=2) + myGgTheme

questionnaires %>% with(cor.test(SPAI, STAI, alternative="greater")) %>% correlation_out()
#questionnaires %>% ggplot(aes(x=SPAI, y=STAI)) + geom_smooth(method="lm", color="black") + geom_point(size=2) + myGgTheme


# Significance of gradients -----------------------------------------------
data.wide %>% pull(Gen_all_lds) %>% t.test(mu = 0, alternative="greater") %>% apa::t_apa(es_ci=T)
#gradients %>% pull(Gen_all_lds) %>% mean() %>% signif(3) %>% paste0("M = ", .)
data.wide %>% summarise(M = mean(Gen_all_lds, na.rm = TRUE),
                        SD = sd(Gen_all_lds, na.rm = TRUE)) %>% tibble()

# Correlation LDS & SPAI/STAI
data.wide %>% with(cor.test(Gen_all_lds, SPAI, alternative="less")) %>% correlation_out()  #apa::cor_apa(r_ci=T)
data.wide %>% with(cor.test(Gen_all_lds, STAI, alternative="less")) %>% correlation_out()

# Correlation Diff & SPAI/STAI
data.wide %>% with(cor.test(Gen_all_diff, SPAI, alternative="less")) %>% correlation_out()  #apa::cor_apa(r_ci=T)
data.wide %>% with(cor.test(Gen_all_diff, STAI, alternative="less")) %>% correlation_out()

# Correlation Level & SPAI/STAI
data.wide %>% with(cor.test(Gen_all_level, SPAI, alternative="greater")) %>% correlation_out()  #apa::cor_apa(r_ci=T)
data.wide %>% with(cor.test(Gen_all_level, STAI, alternative="greater")) %>% correlation_out()

# #Significance of gradients for first half -----------------------------------------------
# data.wide.first.half %>% pull(Gen_all_lds) %>% t.test(mu = 0, alternative="greater") %>% apa::t_apa(es_ci=T)
# #gradients %>% pull(Gen_all_lds) %>% mean() %>% signif(3) %>% paste0("M = ", .)
# data.wide.first.half %>% summarise(M = mean(Gen_all_lds, na.rm = TRUE),
#                         SD = sd(Gen_all_lds, na.rm = TRUE)) %>% tibble()
# 
# # Correlation LDS & SPAI/STAI
# data.wide.first.half %>% with(cor.test(Gen_all_lds, SPAI, alternative="less")) %>% correlation_out()  #apa::cor_apa(r_ci=T)
# data.wide.first.half %>% with(cor.test(Gen_all_lds, STAI, alternative="less")) %>% correlation_out()
# 
# # Correlation Diff & SPAI/STAI
# data.wide.first.half %>% with(cor.test(Gen_all_diff, SPAI, alternative="less")) %>% correlation_out()  #apa::cor_apa(r_ci=T)
# data.wide.first.half %>% with(cor.test(Gen_all_diff, STAI, alternative="less")) %>% correlation_out()
# 
# # Correlation Level & SPAI/STAI
# data.wide.first.half %>% with(cor.test(Gen_all_level, SPAI, alternative="greater")) %>% correlation_out()  #apa::cor_apa(r_ci=T)
# data.wide.first.half %>% with(cor.test(Gen_all_level, STAI, alternative="greater")) %>% correlation_out()


# Replication Anxiety -----------------------------------------------------
#data.wide = read_rds("data.wide.rds" %>% paste0(path.rds, .))
#data.wide %>% select(subject:STAI, Gen_all_dwell:Gen_mn_roiSwitch) %>% View()

#dwell time
data.wide %>% with(cor.test(Gen_eyes_dwell, SPAI, alternative="two.sided")) %>% correlation_out()
data.wide %>% with(cor.test(Gen_mn_dwell, SPAI, alternative="two.sided")) %>% correlation_out()
data.wide %>% with(cor.test(Gen_eyes_dwell.non, SPAI, alternative="two.sided")) %>% correlation_out()
#data.wide %>% with(cor.test(Gen_mn_dwell, Gen_eyes_dwell.non, alternative="less")) %>% correlation_out()
data.wide %>% with(cor.test(Gen_mn_dwell.non, SPAI, alternative="two.sided")) %>% correlation_out()

#this should be practically the same as the "small" eye ANOVA with SPAI as covariate (but results differ!?)
# reg.spai = data.wide %>% gather("dwell.type", "dwell", c("Gen_eyes_dwell", "Gen_mn_dwell", "Gen_eyes_dwell.non", "Gen_mn_dwell.non")) %>% 
#   select(subject:STAI, dwell.type:dwell) %>% 
#   #sample_n(15) %>% #for testing of mutate
#   mutate(#Diagnosticity = ifelse(grepl(".non", dwell.type), "Non-Diagnostic", "Diagnostic") %>% as.factor(),
#          #ROI = ifelse(grepl("_eyes_", dwell.type), "Eyes", "Mouth/Nose") %>% as.factor(),
#          Diagnosticity = ifelse(grepl(".non", dwell.type), -1, 1),
#          ROI = ifelse(grepl("_eyes_", dwell.type), 1, -1),
#          dwell = scale(dwell), SPAI = scale(SPAI)) %>% 
#   lmer(dwell ~ Diagnosticity*ROI*SPAI + (1|subject), .)
#   
# reg.spai %>% summary() %>% print()
# reg.spai %>% lmer.ci()
# #reg.spai %>% lmer.ci(twotailed = F)

#latency
data.wide %>% with(cor.test(Gen_eyes_ms, SPAI, alternative="two.sided")) %>% correlation_out()
data.wide %>% with(cor.test(Gen_mn_ms, SPAI, alternative="two.sided")) %>% correlation_out()
data.wide %>% with(cor.test(Gen_eyes_ms.non, SPAI, alternative="two.sided")) %>% correlation_out()
data.wide %>% with(cor.test(Gen_mn_ms.non, SPAI, alternative="two.sided")) %>% correlation_out()
#data.wide %>% with(cor.test(Gen_all_ms, SPAI, alternative="two.sided")) %>% correlation_out()

# Ratings -----------------------------------------------------------------
#data.wide = read_rds("data.wide.rds" %>% paste0(path.rds, .))

#Ratings: LDS & Square Root of Time to diagnostic ROI
reg.lds.ms.sqr = data.wide %>% filter(subject %in% exclusions.eye.ms == F) %>% #manual exclusion because of extreme latency
  gather("lds.type", "lds", c("Gen_eyes_lds", "Gen_mn_lds")) %>% 
  gather("ms.diag.type", "ms.diag", c("Gen_eyes_ms", "Gen_mn_ms")) %>%
  select(subject:STAI, lds.type:ms.diag) %>% 
  mutate(ms.diag = sqrt(ms.diag),
         lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
         ms.diag.type = ifelse(grepl("_eyes_", ms.diag.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
  filter(lds.type == ms.diag.type) %>% 
  mutate(ms.diag.type = ifelse(ms.diag.type=="Eyes", -1, 1)) %>% 
  mutate(lds = scale(lds), ms.diag = scale(ms.diag), SPAI = scale(SPAI), STAI = scale(STAI)) %>% #z-transform
  lmer(lds ~ ms.diag*ms.diag.type*SPAI + (1|subject), .)
#lmer(lds ~ ms.diag*ms.diag.type + (1|subject), .)

reg.lds.ms.sqr %>% summary() %>% print()
reg.lds.ms.sqr %>% lmer.ci() 
#reg.lds.ms.sqr %>% lmer.ci(twotailed = F)

#ms.diag main effect
data.wide %>% with(cor.test(Gen_all_lds, Gen_all_ms, alternative="less")) %>% apa::cor_apa(r_ci=T) #correlation_out()

print(correl.ms.plot <- data.wide %>% filter(subject %in% exclusions.eye.ms == F) %>% #manual exclusion because of extreme latency
        gather("lds.type", "lds", c("Gen_eyes_lds", "Gen_mn_lds")) %>% 
        gather("ms.diag.type", "ms.diag", c("Gen_eyes_ms", "Gen_mn_ms")) %>%
        select(subject:STAI, lds.type:ms.diag) %>% 
        mutate(lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
               ms.diag.type = ifelse(grepl("_eyes_", ms.diag.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
        filter(lds.type == ms.diag.type) %>% 
        ggplot(aes(x=sqrt(ms.diag), y=lds, color=ms.diag.type, fill=ms.diag.type, shape=ms.diag.type)) +
        geom_smooth(method="lm", size=1.5, alpha = .2) +
        #stat_smooth(method="lm", size=1.5, alpha = .2, fullrange = T) +
        geom_point(size=4, alpha=.8) + 
        ylab("Linear Deviation Score") + xlab(expression("Time to Diagnostic ROI (" * sqrt(ms) * ")")) + labs(color="Diagnostic", fill="Diagnostic", shape="Diagnostic") +
        scale_shape_manual(values=c(16, 15)) +
        myGgTheme +
        theme(
          #aspect.ratio = 1,
          #legend.position = "none",
          legend.position = c(.87, .87))
          #legend.position = c(.5, 1-.87))
)
#ggsave("plots/LDS Ratings x Latency.png", plot=correl.ms.plot, scale=1, device="png", dpi=300, units="in", width=1920/300, height = 1080/300)

dwellers.ms = data.wide %>% mutate(Gen_all_ms.diag.md = median(Gen_all_ms, na.rm=T)) %>% 
  transmute(subject = subject, 
            dwell.type = ifelse(Gen_all_ms < Gen_all_ms.diag.md, "fast", "slow")) %>% na.omit()
#dwellers.ms %>% group_by(dwell.type) %>% summarise(n=n())

correl.ms.plot.gradient = dwellers.ms %>% left_join(ratings.ga.gen.subj, by="subject") %>% na.omit() %>% 
  group_by(threat, dwell.type) %>% summarise(rating.se=se(rating, na.rm=T), rating.m=mean(rating, na.rm=T)) %>% 
  ggplot(aes(x=threat, y=rating.m, color=dwell.type, shape=dwell.type, group=dwell.type)) +
  geom_errorbar(aes(ymin=rating.m-rating.se, ymax=rating.m+rating.se), position=dodge, width=dodge.width) +
  geom_line(position=dodge, size=1) +
  geom_point(position=dodge, size=4.5) +
  ylab("Threat Rating (1-5)") + xlab("Threat") + labs(color="Dwell Type", shape="Dwell Type") +
  scale_color_manual(values=c("chartreuse4", "purple")) +
  theme_bw() + theme(
    #aspect.ratio = 1,
    #legend.position = "none",
    legend.position = c(.9, 1-.83),
    panel.background = element_rect(fill="white", color="white"),
    legend.background = element_rect(fill="white", color="grey"),
    legend.key=element_rect(fill='white'),
    legend.text = element_text(size=14, color="black"),
    legend.title = element_text(size=14, color="black"),
    axis.text = element_text(color="black"),
    axis.text.x = element_text(size=16, color="black"),
    axis.text.y = element_text(size=16, color="black"),
    strip.text.x = element_text(size=12, color="black"),
    axis.ticks.x = element_line(color="black"),
    axis.line.x = element_line(color="black"),
    axis.line.y = element_line(color="black"),
    axis.title = element_text(size=16, color="black"))
print(correl.ms.plot.gradient)
#ggsave("plots/Correl Expectancy LDS x ms (Gradient).png", plot=correl.ms.plot.gradient, scale=1, device="png", dpi=300, units="in", width=1920/300, height = 1080/300)


#ms.diag:ms.diag.type:SPAI
# correl.ms.spaiInteraction = data.wide %>% filter(subject %in% exclusions.eye.ms == F) %>% #manual exclusion because of extreme latency
#   gather("lds.type", "lds", c("Gen_eyes_lds", "Gen_mn_lds")) %>% 
#   gather("ms.diag.type", "ms.diag", c("Gen_eyes_ms", "Gen_mn_ms")) %>%
#   select(subject:SPAI, lds.type:ms.diag) %>% 
#   mutate(lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
#          ms.diag.type = ifelse(grepl("_eyes_", ms.diag.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
#   filter(lds.type == ms.diag.type) %>% 
#   mutate(spai.group = ifelse(SPAI > median(SPAI, na.rm=T), "high", "low") %>% factor(levels=c("low", "high")))
# correl.ms.spaiInteraction %>% summarise(rtest = cor.test(lds, sqrt(ms.diag)) %>% apa::cor_apa(r_ci=T, print=F), #correlation_out(returnString=T), 
#                                         .by=c("spai.group", "ms.diag.type")) %>% arrange(spai.group)
# print(correl.ms.spaiInteraction.plot <- correl.ms.spaiInteraction %>% 
#         ggplot(aes(x=sqrt(ms.diag), y=lds, color=ms.diag.type, fill=ms.diag.type, shape=ms.diag.type)) +
#         facet_wrap(vars(spai.group), scales="free_x", labeller = labeller(spai.group = c(low = "Low Social Anxiety", high = "High Social Anxiety"))) + #labeller = "label_both") +
#         #facet_grid(rows=vars(ms.diag.type), cols=vars(spai.group), scales="free_x", labeller = "label_both") +
#         geom_smooth(method="lm", size=1.5, alpha = .2) +
#         #stat_smooth(method="lm", size=1.5, alpha = .2, fullrange = T) +
#         geom_point(size=4, alpha=.8) + 
#         ylab("Linear Deviation Score") + xlab(expression("Time to Diagnostic ROI (" * sqrt(ms) * ")")) + labs(color="Diagnostic", fill="Diagnostic", shape="Diagnostic") +
#         scale_shape_manual(values=c(16, 15)) +
#         myGgTheme
# )
# #Expected negative correlation between latency & LDS stems from
# #diagnostic eyes       for high anxious subjects
# #diagnostic mouth/nose for low  anxious subjects
# 
# {(correl.ms.plot + theme(axis.title.x=element_blank())) / correl.ms.spaiInteraction.plot + 
#     plot_layout(axis_titles = "collect_x", guides = "collect") + plot_annotation(tag_levels = 'a')} %>% 
#   ggsave("figures/Figure LDS.png", plot=., scale=1.45, device="png", dpi=300, units="in", width=6.5, height = 6.5 / sqrt(2))


#STAI analysis
reg.lds.ms.sqr.stai = data.wide %>% filter(subject %in% exclusions.eye.ms == F) %>% #manual exclusion because of extreme latency
  gather("lds.type", "lds", c("Gen_eyes_lds", "Gen_mn_lds")) %>% 
  gather("ms.diag.type", "ms.diag", c("Gen_eyes_ms", "Gen_mn_ms")) %>%
  select(subject:STAI, lds.type:ms.diag) %>% 
  mutate(ms.diag = sqrt(ms.diag),
         lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
         ms.diag.type = ifelse(grepl("_eyes_", ms.diag.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
  filter(lds.type == ms.diag.type) %>% 
  mutate(ms.diag.type = ifelse(ms.diag.type=="Eyes", -1, 1)) %>% 
  mutate(lds = scale(lds), ms.diag = scale(ms.diag), SPAI = scale(SPAI), STAI = scale(STAI)) %>% #z-transform
  lmer(lds ~ ms.diag*ms.diag.type*STAI + (1|subject), .)

reg.lds.ms.sqr.stai %>% summary() %>% print()
reg.lds.ms.sqr.stai %>% lmer.ci() 
#reg.lds.ms.sqr.stai %>% lmer.ci(twotailed = F)

#STAI main effect
data.wide %>% with(cor.test(Gen_all_lds, STAI, alternative="less")) %>% correlation_out()  #apa::cor_apa(r_ci=T)
data.wide %>% with(ppcor::pcor.test(Gen_all_lds, STAI, SPAI, method="pearson")) %>% 
  rename(r = estimate, p = p.value, t = statistic) %>% 
  mutate(p_onesided = if_else(r < 0, p/2, 1-p/2)) %>% rename(p_twosided = p) #TODO CI? => do by hand with residuals?

print(correl.ms.stai.plot <- data.wide %>% filter(subject %in% exclusions.eye.ms == F) %>% #manual exclusion because of extreme latency
        gather("lds.type", "lds", c("Gen_eyes_lds", "Gen_mn_lds")) %>% 
        gather("ms.diag.type", "ms.diag", c("Gen_eyes_ms", "Gen_mn_ms")) %>%
        select(subject:STAI, lds.type:ms.diag) %>% 
        mutate(lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
               ms.diag.type = ifelse(grepl("_eyes_", ms.diag.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
        filter(lds.type == ms.diag.type) %>% 
        ggplot(aes(x=STAI, y=lds, color=ms.diag.type, fill=ms.diag.type, shape=ms.diag.type)) +
        geom_smooth(method="lm", size=1.5, alpha = .2) +
        #stat_smooth(method="lm", size=1.5, alpha = .2, fullrange = T) +
        geom_point(size=4, alpha=.8) + 
        ylab("Linear Deviation Score") + labs(color="Diagnostic", fill="Diagnostic", shape="Diagnostic") +
        scale_shape_manual(values=c(16, 15)) +
        myGgTheme +
        theme(
          #aspect.ratio = 1,
          #legend.position = "none",
          legend.position = c(.87, 1-.87))
      #legend.position = c(.5, 1-.87))
)

#ms.diag:ms.diag.type:STAI
correl.ms.staiInteraction = data.wide %>% filter(subject %in% exclusions.eye.ms == F) %>% #manual exclusion because of extreme latency
  gather("lds.type", "lds", c("Gen_eyes_lds", "Gen_mn_lds")) %>% 
  gather("ms.diag.type", "ms.diag", c("Gen_eyes_ms", "Gen_mn_ms")) %>%
  select(subject:STAI, lds.type:ms.diag) %>% 
  mutate(lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
         ms.diag.type = ifelse(grepl("_eyes_", ms.diag.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
  filter(lds.type == ms.diag.type) %>% 
  mutate(stai.group = ifelse(STAI > median(STAI, na.rm=T), "high", "low") %>% factor(levels=c("low", "high")))
correl.ms.staiInteraction %>% summarise(rtest = cor.test(lds, sqrt(ms.diag)) %>% correlation_out(returnString=T), 
                                        .by=c("stai.group", "ms.diag.type")) %>% arrange(stai.group)
print(correl.ms.staiInteraction.plot <- correl.ms.staiInteraction %>% 
        ggplot(aes(x=sqrt(ms.diag), y=lds, color=ms.diag.type, fill=ms.diag.type, shape=ms.diag.type)) +
        facet_wrap(vars(stai.group), scales="free_x", labeller = "label_both") +
        #facet_grid(rows=vars(ms.diag.type), cols=vars(stai.group), scales="free_x", labeller = "label_both") +
        geom_smooth(method="lm", size=1.5, alpha = .2) +
        #stat_smooth(method="lm", size=1.5, alpha = .2, fullrange = T) +
        geom_point(size=4, alpha=.8) + 
        ylab("Linear Deviation Score") + labs(color="Diagnostic", fill="Diagnostic", shape="Diagnostic") +
        scale_shape_manual(values=c(16, 15)) +
        myGgTheme
)

#Expected negative correlation between latency & LDS stems from diagnostic eyes for high anxious subjects

#interaction (three-way)
correl.ms.stai.diag.Interaction = data.wide %>% filter(subject %in% exclusions.eye.ms == F) %>% #manual exclusion because of extreme latency
  gather("lds.type", "lds", c("Gen_eyes_lds", "Gen_mn_lds")) %>% 
  gather("ms.diag.type", "ms.diag", c("Gen_eyes_ms", "Gen_mn_ms")) %>%
  select(subject:STAI, lds.type:ms.diag) %>% 
  mutate(lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
         ms.diag.type = ifelse(grepl("_eyes_", ms.diag.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
  filter(lds.type == ms.diag.type) %>% 
  mutate(stai.group = ifelse(STAI > median(STAI, na.rm=T), "high", "low") %>% factor(levels=c("low", "high")))
correl.ms.stai.diag.Interaction %>% filter(stai.group == "high") %>% summarise(rtest = cor.test(lds, ms.diag) %>% correlation_out(returnString=T), 
                                        .by=c("ms.diag.type"))
correl.ms.stai.diag.Interaction %>% filter(stai.group == "low") %>% summarise(rtest = cor.test(lds, ms.diag) %>% correlation_out(returnString=T), 
                                              .by=c("ms.diag.type"))
print(correl.ms.stai.diag.Interaction.plot <- correl.ms.stai.diag.Interaction %>% 
        ggplot(aes(x=ms.diag, y=lds, color=ms.diag.type, fill=ms.diag.type, shape=ms.diag.type)) +
        facet_wrap(vars(stai.group), scales="free_x") +
        #facet_grid(rows=vars(ms.diag.type), cols=vars(stai.group), scales="free_x") +
        geom_smooth(method="lm", size=1.5, alpha = .2) +
        #stat_smooth(method="lm", size=1.5, alpha = .2, fullrange = T) +
        geom_point(size=4, alpha=.8) + 
        ylab("Linear Deviation Score") + labs(color="Diagnostic", fill="Diagnostic", shape="Diagnostic") +
        scale_shape_manual(values=c(16, 15)) +
        myGgTheme
)


#interaction SPAI x diagnostic region
correl.ms.spaiInteraction = data.wide %>% filter(subject %in% exclusions.eye.ms == F) %>% #manual exclusion because of extreme latency
  gather("lds.type", "lds", c("Gen_eyes_lds", "Gen_mn_lds")) %>% 
  gather("ms.diag.type", "ms.diag", c("Gen_eyes_ms", "Gen_mn_ms")) %>%
  select(subject:SPAI, lds.type:ms.diag) %>% 
  mutate(lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
         ms.diag.type = ifelse(grepl("_eyes_", ms.diag.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
  filter(lds.type == ms.diag.type) %>% 
  mutate(spai.group = ifelse(SPAI > median(SPAI, na.rm=T), "high", "low") %>% factor(levels=c("low", "high")))
correl.ms.spaiInteraction %>% summarise(rtest = cor.test(lds, SPAI) %>% correlation_out(returnString=T), 
                                        .by=c("ms.diag.type", "spai.group"))
print(correl.ms.spaiInteraction.plot <- correl.ms.spaiInteraction %>% 
        ggplot(aes(x=SPAI, y=lds, color=ms.diag.type, fill=ms.diag.type, shape=ms.diag.type)) +
        facet_wrap(vars(spai.group), scales="free_x") +
        #facet_grid(rows=vars(ms.diag.type), cols=vars(stai.group), scales="free_x") +
        geom_smooth(method="lm", size=1.5, alpha = .2) +
        #stat_smooth(method="lm", size=1.5, alpha = .2, fullrange = T) +
        geom_point(size=4, alpha=.8) + 
        ylab("Linear Deviation Score") + labs(color="Diagnostic", fill="Diagnostic", shape="Diagnostic") +
        scale_shape_manual(values=c(16, 15)) +
        myGgTheme
)

#interaction SPAI (three-way)
correl.ms.spai.diag.Interaction = data.wide %>% filter(subject %in% exclusions.eye.ms == F) %>% #manual exclusion because of extreme latency
  gather("lds.type", "lds", c("Gen_eyes_lds", "Gen_mn_lds")) %>% 
  gather("ms.diag.type", "ms.diag", c("Gen_eyes_ms", "Gen_mn_ms")) %>%
  select(subject:SPAI, lds.type:ms.diag) %>% 
  mutate(lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
         ms.diag.type = ifelse(grepl("_eyes_", ms.diag.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
  filter(lds.type == ms.diag.type) %>% 
  mutate(spai.group = ifelse(SPAI > median(SPAI, na.rm=T), "high", "low") %>% factor(levels=c("low", "high")))
correl.ms.spai.diag.Interaction %>% filter(spai.group == "high") %>% summarise(rtest = cor.test(lds, ms.diag) %>% correlation_out(returnString=T), 
                                                                               .by=c("ms.diag.type"))
correl.ms.spai.diag.Interaction %>% filter(spai.group == "low") %>% summarise(rtest = cor.test(lds, ms.diag) %>% correlation_out(returnString=T), 
                                                                              .by=c("ms.diag.type"))
print(correl.ms.spai.diag.Interaction.plot <- correl.ms.spai.diag.Interaction %>% 
        ggplot(aes(x=ms.diag, y=lds, color=ms.diag.type, fill=ms.diag.type, shape=ms.diag.type)) +
        facet_wrap(vars(spai.group), scales="free_x") +
        #facet_grid(rows=vars(ms.diag.type), cols=vars(stai.group), scales="free_x") +
        geom_smooth(method="lm", size=1.5, alpha = .2) +
        #stat_smooth(method="lm", size=1.5, alpha = .2, fullrange = T) +
        geom_point(size=4, alpha=.8) + 
        ylab("Linear Deviation Score") + labs(color="Diagnostic", fill="Diagnostic", shape="Diagnostic") +
        scale_shape_manual(values=c(16, 15)) +
        myGgTheme
)

# #Ratings: LDS & Square Root of Time to non-diagnostic ROI ---- CHECK LMM!!
# 
# reg.lds.ms.non.sqr = data.wide %>% filter(subject %in% exclusions.eye.ms == F) %>% #manual exclusion because of extreme latency
#   gather("lds.type", "lds", c("Gen_eyes_lds", "Gen_mn_lds")) %>% 
#   gather("ms.diag.type", "ms.non", c("Gen_eyes_ms.non", "Gen_mn_ms.non")) %>%
#   select(subject:STAI, lds.type:ms.non) %>% 
#   mutate(ms.non = sqrt(ms.non),
#          lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
#          ms.diag.type = ifelse(grepl("_eyes_", ms.diag.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
#   filter(lds.type == ms.diag.type) %>% 
#   mutate(ms.diag.type = ifelse(ms.diag.type=="Eyes", -1, 1)) %>% 
#   mutate(lds = scale(lds), ms.diag = scale(ms.non), SPAI = scale(SPAI), STAI = scale(STAI)) %>% #z-transform
#   #lmer(lds ~ ms.diag*ms.diag.type*SPAI + (1|subject), .)
#   lmer(lds ~ ms.non*ms.diag.type*STAI + (1|subject), .)
# #lmer(lds ~ ms.diag*ms.diag.type + (1|subject), .)
# 
# reg.lds.ms.non.sqr %>% summary() %>% print()
# reg.lds.ms.non.sqr %>% lmer.ci() 
# #reg.lds.ms.sqr %>% lmer.ci(twotailed = F)
# 
# #ms.diag main effect
# data.wide %>% with(cor.test(Gen_all_lds, Gen_all_ms.non, alternative="less")) %>% correlation_out()  #apa::cor_apa(r_ci=T)
# 
# print(correl.ms.non.plot <- data.wide %>% filter(subject %in% exclusions.eye.ms == F) %>% #manual exclusion because of extreme latency
#         gather("lds.type", "lds", c("Gen_eyes_lds", "Gen_mn_lds")) %>% 
#         gather("ms.diag.type", "ms.non", c("Gen_eyes_ms.non", "Gen_mn_ms.non")) %>%
#         select(subject:STAI, lds.type:ms.non) %>% 
#         mutate(lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
#                ms.diag.type = ifelse(grepl("_eyes_", ms.diag.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
#         filter(lds.type == ms.diag.type) %>% 
#         ggplot(aes(x=sqrt(ms.non), y=lds, color=ms.diag.type, fill=ms.diag.type, shape=ms.diag.type)) +
#         geom_smooth(method="lm", size=1.5, alpha = .2) +
#         #stat_smooth(method="lm", size=1.5, alpha = .2, fullrange = T) +
#         geom_point(size=4, alpha=.8) + 
#         ylab("Linear Deviation Score") + xlab(expression("Time to Non-Diagnostic ROI (" * sqrt(ms) * ")")) + labs(color="Diagnostic", fill="Diagnostic", shape="Diagnostic") +
#         scale_shape_manual(values=c(16, 15)) +
#         myGgTheme +
#         theme(
#           #aspect.ratio = 1,
#           #legend.position = "none",
#           legend.position = c(.87, 1-.87))
#       #legend.position = c(.5, 1-.87))
# )

# #Ratings: Level & Square Root of Time to diagnostic ROI
# reg.level.ms.sqr = data.wide %>% filter(subject %in% exclusions.eye.ms == F) %>% #manual exclusion because of extreme latency
#   gather("level.type", "level", c("Gen_eyes_level", "Gen_mn_level")) %>% 
#   gather("ms.diag.type", "ms.diag", c("Gen_eyes_ms", "Gen_mn_ms")) %>%
#   select(subject:STAI, level.type:ms.diag) %>% 
#   mutate(ms.diag = sqrt(ms.diag),
#          level.type = ifelse(grepl("_eyes_", level.type), "Eyes", "Mouth/Nose") %>% as.factor(),
#          ms.diag.type = ifelse(grepl("_eyes_", ms.diag.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
#   filter(level.type == ms.diag.type) %>% 
#   mutate(ms.diag.type = ifelse(ms.diag.type=="Eyes", -1, 1)) %>% 
#   mutate(level = scale(level), ms.diag = scale(ms.diag), SPAI = scale(SPAI), STAI = scale(STAI)) %>% #z-transform
#   #lmer(level ~ ms.diag*ms.diag.type*SPAI + (1|subject), .)
#   lmer(level ~ ms.diag*ms.diag.type*STAI + (1|subject), .)
# #lmer(lds ~ ms.diag*ms.diag.type + (1|subject), .)
# 
# reg.level.ms.sqr %>% summary() %>% print()
# reg.level.ms.sqr %>% lmer.ci() 
# #reg.lds.ms.sqr %>% lmer.ci(twotailed = F)
# 
# #ms.diag main effect
# data.wide %>% with(cor.test(Gen_all_level, Gen_all_ms, alternative="less")) %>% correlation_out()  #apa::cor_apa(r_ci=T)
# 
# print(correl.level.ms.plot <- data.wide %>% filter(subject %in% exclusions.eye.ms == F) %>% #manual exclusion because of extreme latency
#         gather("level.type", "level", c("Gen_eyes_level", "Gen_mn_level")) %>% 
#         gather("ms.diag.type", "ms.diag", c("Gen_eyes_ms", "Gen_mn_ms")) %>%
#         select(subject:STAI, level.type:ms.diag) %>% 
#         mutate(level.type = ifelse(grepl("_eyes_", level.type), "Eyes", "Mouth/Nose") %>% as.factor(),
#                ms.diag.type = ifelse(grepl("_eyes_", ms.diag.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
#         filter(level.type == ms.diag.type) %>% 
#         ggplot(aes(x=sqrt(ms.diag), y=level, color=ms.diag.type, fill=ms.diag.type, shape=ms.diag.type)) +
#         geom_smooth(method="lm", size=1.5, alpha = .2) +
#         #stat_smooth(method="lm", size=1.5, alpha = .2, fullrange = T) +
#         geom_point(size=4, alpha=.8) + 
#         ylab("Ratings Level") + xlab(expression("Time to Diagnostic ROI (" * sqrt(ms) * ")")) + labs(color="Diagnostic", fill="Diagnostic", shape="Diagnostic") +
#         scale_shape_manual(values=c(16, 15)) +
#         myGgTheme +
#         theme(
#           #aspect.ratio = 1,
#           #legend.position = "none",
#           legend.position = c(.87, 1-.87))
#       #legend.position = c(.5, 1-.87))
# )

#Ratings: LDS & Time to diagnostic ROI (check that sqrt did not totally skew results)
reg.lds.ms = data.wide %>% filter(subject %in% exclusions.eye.ms == F) %>% #manual exclusion because of extreme latency
  gather("lds.type", "lds", c("Gen_eyes_lds", "Gen_mn_lds")) %>% 
  gather("ms.diag.type", "ms.diag", c("Gen_eyes_ms", "Gen_mn_ms")) %>%
  select(subject:STAI, lds.type:ms.diag) %>% 
  mutate(lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
         ms.diag.type = ifelse(grepl("_eyes_", ms.diag.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
  filter(lds.type == ms.diag.type) %>% 
  mutate(ms.diag.type = ifelse(ms.diag.type=="Eyes", -1, 1)) %>% 
  #mutate(lds = scale(lds), ms.diag = scale(ms.diag), SPAI = scale(SPAI), STAI = scale(STAI)) %>% #z-transform
  #lmer(lds ~ ms.diag*ms.diag.type*SPAI + (1|subject), .)
  #lmer(lds ~ ms.diag*ms.diag.type*STAI + (1|subject), .)
  lmer(lds ~ ms.diag*ms.diag.type + (1|subject), .)

reg.lds.ms %>% summary() %>% print()
reg.lds.ms %>% lmer.ci() 
#reg.lds.ms %>% lmer.ci(twotailed = F)

data.wide %>% filter(subject %in% exclusions.eye.ms == F) %>% #manual exclusion because of extreme latency
  gather("lds.type", "lds", c("Gen_eyes_lds", "Gen_mn_lds")) %>% 
  gather("ms.diag.type", "ms.diag", c("Gen_eyes_ms", "Gen_mn_ms")) %>%
  select(subject:STAI, lds.type:ms.diag) %>% 
  mutate(lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
         ms.diag.type = ifelse(grepl("_eyes_", ms.diag.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
  filter(lds.type == ms.diag.type) %>% 
  ggplot(aes(x=ms.diag, y=lds, color=ms.diag.type, fill=ms.diag.type, shape=ms.diag.type)) +
  geom_smooth(method="lm", size=1.5, alpha = .2) +
  #stat_smooth(method="lm", size=1.5, alpha = .2, fullrange = T) +
  geom_point(size=4, alpha=.8) + 
  ylab("Linear Deviation Score") + xlab("Time to Diagnostic ROI (ms)") + labs(color="Diagnostic", fill="Diagnostic", shape="Diagnostic") +
  scale_shape_manual(values=c(16, 15)) +
  myGgTheme +
  theme(
    #aspect.ratio = 1,
    #legend.position = "none",
    legend.position = c(.87, 1-.87))
    #legend.position = c(.5, 1-.87))



#Ratings: LDS & Fixations
reg.lds.dwell = data.wide %>% filter(subject %in% exclusions.eye.dwell == F) %>% #manual exclusion because of extreme dwell
  gather("lds.type", "lds", c("Gen_eyes_lds", "Gen_mn_lds")) %>% 
  gather("dwell.type", "dwell", c("Gen_eyes_dwell", "Gen_mn_dwell")) %>%
  select(subject:STAI, lds.type:dwell) %>% 
  mutate(lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
         dwell.type = ifelse(grepl("_eyes_", dwell.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
  filter(lds.type == dwell.type) %>% 
  #mutate(dwell.type = factor(dwell.type)) %>% #dummy coding
  mutate(dwell.type = ifelse(dwell.type=="Eyes", -1, 1)) %>% #effect coding
  mutate(lds = scale(lds), dwell = scale(dwell), SPAI = scale(SPAI), STAI = scale(STAI)) %>% #z-transform
  lmer(lds ~ dwell*dwell.type*SPAI + (1|subject), .)
  #lmer(lds ~ dwell*dwell.type*STAI + (1|subject), .)
  #lmer(lds ~ dwell*dwell.type + (1|subject), .)

reg.lds.dwell %>% summary() %>% print()
reg.lds.dwell %>% lmer.ci()
#reg.lds.dwell %>% lmer.ci(twotailed = F)

dwellers = data.wide %>% mutate(Gen_all_dwell.md = median(Gen_all_dwell, na.rm=T)) %>% 
  transmute(subject = subject, 
            dwell.type = ifelse(Gen_all_dwell > Gen_all_dwell.md, "high", "low")) %>% na.omit()
#dwellers %>% group_by(dwell.type) %>% summarise(n=n())

correl.dwell.plot.gradient = dwellers %>% left_join(ratings.ga.gen.subj, by="subject") %>% na.omit() %>% 
  group_by(threat, dwell.type) %>% summarise(rating.se=se(rating, na.rm=T), rating.m=mean(rating, na.rm=T)) %>% 
  ggplot(aes(x=threat, y=rating.m, color=dwell.type, shape=dwell.type, group=dwell.type)) +
  #geom_errorbar(aes(ymin=rating.m-rating.se*1.96, ymax=rating.m+rating.se*1.96), position=dodge, width=dodge.width) +
  geom_errorbar(aes(ymin=rating.m-rating.se, ymax=rating.m+rating.se), position=dodge, width=dodge.width) +
  geom_line(position=dodge, size=1) +
  geom_point(position=dodge, size=4.5) +
  ylab("Threat Rating (1-5)") + xlab("Threat") + labs(color="Dwell Type", shape="Dwell Type") +
  scale_color_manual(values=c("chartreuse4", "purple")) +
  theme_bw() + theme(
    #aspect.ratio = 1,
    #legend.position = "none",
    #legend.position = c(.9, 1-.87),
    legend.position = c(.9, 1-.83),
    panel.background = element_rect(fill="white", color="white"),
    legend.background = element_rect(fill="white", color="grey"),
    legend.key=element_rect(fill='white'),
    legend.text = element_text(size=14, color="black"),
    legend.title = element_text(size=14, color="black"),
    axis.text = element_text(color="black"),
    axis.text.x = element_text(size=16, color="black"),
    axis.text.y = element_text(size=16, color="black"),
    strip.text.x = element_text(size=12, color="black"),
    axis.ticks.x = element_line(color="black"),
    axis.line.x = element_line(color="black"),
    axis.line.y = element_line(color="black"),
    axis.title = element_text(size=16, color="black"))
print(correl.dwell.plot.gradient)
#ggsave("plots/Correl Expectancy LDS x Dwell (Gradient).png", plot=correl.dwell.plot.gradient, scale=1, device="png", dpi=300, units="in", width=1920/300, height = 1080/300)


#dwell*dwell.type
# print(correl.dwell.plot <- data.wide %>% filter(subject %in% exclusions.eye.dwell == F) %>% #manual exclusion because of extreme dwell
#         gather("lds.type", "lds", c("Gen_eyes_lds", "Gen_mn_lds")) %>% 
#         gather("dwell.type", "dwell", c("Gen_eyes_dwell", "Gen_mn_dwell")) %>%
#         select(subject:STAI, lds.type:dwell) %>% 
#         mutate(lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
#                dwell.type = ifelse(grepl("_eyes_", dwell.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
#         filter(lds.type == dwell.type) %>% 
#         ggplot(aes(x=dwell, y=lds, color=dwell.type, fill=dwell.type, shape=dwell.type)) +
#         #geom_line(aes(group=subject), color="black", linetype="dashed") + 
#         geom_smooth(method="lm", size=1.5, alpha = .2) +
#         geom_point(size=4, alpha=.8) + 
#         ylab("Linear Deviation Score") + xlab("Diagnostic Dwell (%)") + labs(color="Diagnostic", fill="Diagnostic", shape="Diagnostic") +
#         scale_shape_manual(values=c(16, 15)) +
#         myGgTheme +
#         theme(
#           #legend.position = "none",
#           #legend.position = c(.87, .87))
#           #legend.position = c(1-.87, 1-.87))
#           legend.position = c(1-.87, .87))
# )

#STAI*dwell.type
# print(correl.stai.plot <- data.wide %>% gather("lds.type", "lds", c("Gen_eyes_lds", "Gen_mn_lds")) %>% 
#         gather("dwell.type", "dwell", c("Gen_eyes_dwell", "Gen_mn_dwell")) %>%
#         select(subject:STAI, lds.type:dwell) %>% 
#         mutate(lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
#                dwell.type = ifelse(grepl("_eyes_", dwell.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
#         filter(lds.type == dwell.type) %>% 
#         ggplot(aes(x=STAI, y=lds, color=dwell.type, fill=dwell.type)) +
#         #geom_line(aes(group=subject), color="black", linetype="dashed") + 
#         geom_smooth(method="lm", size=1.5, alpha = .2) +
#         geom_point(size=2) +
#         ylab("Linear Deviation Score") + labs(color="Diagnostic", fill="Diagnostic") +
#         myGgTheme +
#         theme(
#           #legend.position = "none",
#           #legend.position = c(.87, .87),
#           legend.position = c(1-.87, 1-.87),
#         )
# )
# data.wide %>% with(cor.test(Gen_mn_lds, STAI, alternative="two.sided")) %>% correlation_out()
# data.wide %>% with(cor.test(Gen_eyes_lds, STAI, alternative="two.sided")) %>% correlation_out()

#cowplot::plot_grid(correl.dwell.plot, correl.ms.plot, ncol=1, labels="auto") #Figure 6 (preprint) or 5 (manuscript)


# Pupil (no effects)-----------------------------------------------------
#Pupil: LDS & Dwell
reg.pup.dwell = data.wide %>% gather("lds.type", "lds", c("Pup_Gen_eyes_lds", "Pup_Gen_mn_lds")) %>% 
  gather("dwell.type", "dwell", c("Gen_eyes_dwell", "Gen_mn_dwell")) %>%
  select(subject:STAI, lds.type:dwell) %>% 
  mutate(lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
         dwell.type = ifelse(grepl("_eyes_", dwell.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
  filter(lds.type == dwell.type) %>% 
  #mutate(dwell.type = factor(dwell.type)) %>% 
  mutate(dwell.type = ifelse(dwell.type=="Eyes", -1, 1)) %>% 
  mutate(lds = scale(lds), dwell = scale(dwell), SPAI = scale(SPAI), STAI = scale(STAI)) %>% #z-transform
  lmer(lds ~ dwell*dwell.type*SPAI + (1|subject), .) 
  #lmer(lds ~ dwell*dwell.type*STAI + (1|subject), .) 
  #lmer(lds ~ dwell*dwell.type + (1|subject), .) 

reg.pup.dwell %>% summary() %>% print()
reg.pup.dwell %>% lmer.ci()
#reg.pup.dwell %>% lmer.ci(twotailed = F)

data.wide %>% gather("lds.type", "lds", c("Pup_Gen_eyes_lds", "Pup_Gen_mn_lds")) %>% 
  gather("dwell.type", "dwell", c("Gen_eyes_dwell", "Gen_mn_dwell")) %>%
  select(lds.type:dwell) %>% 
  mutate(lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
         dwell.type = ifelse(grepl("_eyes_", dwell.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
  filter(lds.type == dwell.type) %>% 
  ggplot(aes(x=dwell, y=lds, color=dwell.type, fill=dwell.type)) +
  geom_smooth(method="lm", size=1.5, alpha = .2) +
  geom_point(size=2) +
  ylab("Linear Deviation Score (Pupil)") + xlab("Diagnostic Dwell (%)") + labs(color="Diagnostic", fill="Diagnostic") +
  myGgTheme + theme(
    #legend.position = "none",
    legend.position = c(1-.87, .87)
  )


#Pupil: LDS & Square Root of Time to diagnostic ROI
reg.pup.ms = data.wide %>% gather("lds.type", "lds", c("Pup_Gen_eyes_lds", "Pup_Gen_mn_lds")) %>% 
  gather("ms.diag.type", "ms.diag", c("Gen_eyes_ms", "Gen_mn_ms")) %>%
  select(subject:STAI, lds.type:ms.diag) %>% 
  mutate(ms.diag = sqrt(ms.diag),
         lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
         ms.diag.type = ifelse(grepl("_eyes_", ms.diag.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
  na.omit() %>% 
  filter(lds.type == ms.diag.type) %>% 
  mutate(ms.diag.type = ifelse(ms.diag.type=="Eyes", -1, 1)) %>% 
  mutate(lds = scale(lds), ms.diag = scale(ms.diag), SPAI = scale(SPAI), STAI = scale(STAI)) %>% #z-transform
  lmer(lds ~ ms.diag*ms.diag.type*SPAI + (1|subject), .) 
  #lmer(lds ~ ms.diag*ms.diag.type*STAI + (1|subject), .) 
  #lmer(lds ~ ms.diag*ms.diag.type + (1|subject), .) 

reg.pup.ms %>% summary() %>% print()
reg.pup.ms %>% lmer.ci()
reg.pup.ms %>% lmer.ci(twotailed = F)

data.wide %>% gather("lds.type", "lds", c("Pup_Gen_eyes_lds", "Pup_Gen_mn_lds")) %>% 
  gather("ms.diag.type", "ms.diag", c("Gen_eyes_ms", "Gen_mn_ms")) %>%
  select(lds.type:ms.diag) %>% 
  mutate(lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
         ms.diag.type = ifelse(grepl("_eyes_", ms.diag.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
  filter(lds.type == ms.diag.type) %>% 
  ggplot(aes(x=sqrt(ms.diag), y=lds, color=ms.diag.type, fill=ms.diag.type)) +
  geom_smooth(method="lm", size=1.5, alpha = .2) +
  #stat_smooth(method="lm", size=1.5, alpha = .2, fullrange = T) +
  geom_point(size=2) +
  ylab("Linear Deviation Score (Pupil)") + xlab(expression("Time to Diagnostic ROI (" * sqrt(ms) * ")")) + labs(color="Diagnostic", fill="Diagnostic") +
  myGgTheme + theme(
    #legend.position = "none",
    legend.position = c(.87, .87))


# Heartrate -----------------------------------------------------------------
#Heartrate: LDS & Fixations
heart.dwell = data.wide %>% filter(subject %in% exclusions.eye.dwell == F) %>% #manual exclusion because of extreme dwell
  gather("lds.type", "lds", c("HR_Gen_eyes_lds", "HR_Gen_mn_lds")) %>% 
  gather("dwell.type", "dwell", c("Gen_eyes_dwell", "Gen_mn_dwell")) %>%
  select(subject:STAI, lds.type:dwell) %>% 
  mutate(lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
         dwell.type = ifelse(grepl("_eyes_", dwell.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
  filter(lds.type == dwell.type) %>% 
  mutate(dwell.type.dummy = ifelse(dwell.type=="Eyes", -1, 1)) %>% 
  mutate(lds.z = scale(lds), dwell.z = scale(dwell), SPAI.z = scale(SPAI), STAI.z = scale(STAI)) #z-transform
reg.heart.dwell = heart.dwell %>% 
  lmer(lds.z ~ dwell.z*dwell.type.dummy*SPAI.z + (1|subject), .)
  #lmer(lds.z ~ dwell.z*dwell.type.dummy*STAI.z + (1|subject), .)
  #lmer(lds.z ~ dwell.z*dwell.type.dummy + (1|subject), .)

reg.heart.dwell %>% summary()
reg.heart.dwell %>% lmer.ci() 
#reg.heart.dwell %>% lmer.ci(twotailed = F)

#main effect dwell (marginal)
print(reg.dwell.hr.plot <- heart.dwell %>%
        ggplot(aes(x=dwell, y=lds, color=dwell.type, fill=dwell.type, shape=dwell.type)) +
        #geom_line(aes(group=subject), color="black", linetype="dashed") +
        geom_smooth(method="lm", size=1.5, alpha = .2) +
        geom_point(size=4, alpha=.8) +
        ylab("LDS of Heart Change") + xlab("Diagnostic Dwell (%)") + labs(color="Diagnostic", fill="Diagnostic", shape="Diagnostic") +
        scale_shape_manual(values=c(16, 15)) +
        myGgTheme +
        theme(
          #legend.position = "none",
          #legend.position = c(.87, .87),
          legend.position = c(.87, 1-.87),
        )
)

#dwell.type x SPAI
heart.dwell %>% group_by(dwell.type) %>% summarise(cor.test(lds, SPAI) %>% apa::cor_apa(r_ci=T, print=F))
heart.dwell %>% ggplot(aes(x=SPAI, y=lds, color=dwell.type, fill=dwell.type, shape=dwell.type)) +
  facet_wrap(vars(dwell.type)) +
  geom_smooth(method="lm", size=1.5, color="black", alpha = .2) +
  #geom_point(size=4, alpha=.8) +
  geom_point(size=4) + 
  ylab("LDS of Heart Change") + xlab("SPAI") + labs(color="Diagnostic", fill="Diagnostic", shape="Diagnostic") +
  scale_shape_manual(values=c(16, 15)) +
  scale_color_viridis_d() + scale_fill_viridis_d() +
  myGgTheme +
  theme(
    #legend.position = "none",
    #legend.position = c(.87, .87),
    #legend.position = c(.87, 1-.87),
  )

#dwell x SPAI
SDs = -2:2
spai.descr = with(questionnaires, data.frame(SPAI=mean(SPAI)+sd(SPAI)*SDs, SPAI.z=SDs))
reg.heart.spaiSD.predict = spai.descr %>%
  expand.grid.df(data.frame(dwell=unique(data.wide$Gen_all_dwell %>% na.omit())) %>% mutate(dwell.z=scale(dwell))) %>%
  expand.grid.df(data.frame(subject=0, dwell.type.dummy=0)) %>% #include mockup subject for population-level prediction (could avoid this and add "re.form=NA" to predict but less flexibility with additional subject-level predictions or based on non-lmer model objects, i.e. predict.merMod)
  mutate(lds = predict(reg.heart.dwell, newdata=., allow.new.levels=T)) #predict values from dwell time & z-scaled SPAI

print(reg.dwell.hr.spai.plot <- heart.dwell %>% group_by(subject) %>%
        summarise(dwell=mean(dwell, na.rm=T), lds=mean(lds, na.rm=T), SPAI=mean(SPAI, na.rm=T)) %>%
        ggplot(aes(x=dwell, y=lds, color=SPAI, group=SPAI)) +
        geom_hline(yintercept = 0, linetype="dashed") +
        geom_line(data=reg.heart.spaiSD.predict, size=4) +
        geom_point(size=2) +
        ylab("LDS of Heart Change") + xlab("Diagnostic Dwell (%)") +
        myGgTheme + scale_color_viridis_c()
)


#Heartrate: LDS & Square Root of Time to diagnostic ROI
heart.ms = data.wide %>% filter(subject %in% exclusions.eye.ms == F) %>% #manual exclusion because of extreme latency
  gather("lds.type", "lds", c("HR_Gen_eyes_lds", "HR_Gen_mn_lds")) %>% 
  gather("ms.diag.type", "ms.diag", c("Gen_eyes_ms", "Gen_mn_ms")) %>%
  select(subject:STAI, lds.type:ms.diag) %>% 
  mutate(ms.diag = sqrt(ms.diag),
         lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
         ms.diag.type = ifelse(grepl("_eyes_", ms.diag.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
  filter(lds.type == ms.diag.type) %>% 
  mutate(ms.diag.type = ifelse(ms.diag.type=="Eyes", -1, 1)) %>% 
  mutate(lds.z = scale(lds), ms.diag.z = scale(ms.diag), SPAI.z = scale(SPAI), STAI.z = scale(STAI)) #z-transform
reg.heart.ms = heart.ms %>% 
  lmer(lds.z ~ ms.diag.z*ms.diag.type*SPAI.z + (1|subject), .)
  #lmer(lds.z ~ ms.diag.z*ms.diag.type*STAI.z + (1|subject), .)
  #lmer(lds.z ~ ms.diag.z*ms.diag.type + (1|subject), .)

reg.heart.ms %>% summary() %>% print()
reg.heart.ms %>% lmer.ci()
#reg.heart.ms %>% lmer.ci(twotailed = F)

#SPAI x ms.diag (sqrt)
SDs = -1:1 #-2:2
spai.descr = with(questionnaires, data.frame(SPAI=mean(SPAI)+sd(SPAI)*SDs, SPAI.z=SDs))
reg.heart.ms.spaiSD.predict = spai.descr %>%
  expand.grid.df(data.frame(ms.diag=unique(data.wide$Gen_all_ms %>% na.omit() %>% sqrt())) %>% mutate(ms.diag.z=scale(ms.diag))) %>%
  expand.grid.df(data.frame(subject=0, ms.diag.type=0)) %>% #include mockup subject for population-level prediction (could avoid this and add "re.form=NA" to predict but less flexibility with additional subject-level predictions or based on non-lmer model objects, i.e. predict.merMod)
  mutate(lds = predict(reg.heart.ms, newdata=., allow.new.levels=T)) #predict values from dwell time & z-scaled SPAI

print(reg.ms.hr.spai.plot <- heart.ms %>% group_by(subject) %>% 
        summarise(ms.diag=mean(ms.diag, na.rm=T), lds=mean(lds, na.rm=T), SPAI=mean(SPAI, na.rm=T)) %>% 
        ggplot(aes(x=ms.diag, y=lds, color=SPAI, group=SPAI)) +
        geom_line(data=reg.heart.ms.spaiSD.predict, size=4) +
        geom_hline(yintercept = 0, linetype="dashed") +
        geom_point(size=2) +
        ylab("LDS of Heart Change") + xlab(expression("Time to Diagnostic ROI (" * sqrt(ms) * ")")) + 
        myGgTheme + scale_color_viridis_c()
)

heart.ms %>% summarise(.by=subject, ms.diag=mean(ms.diag, na.rm=T), lds=mean(lds, na.rm=T), SPAI=mean(SPAI, na.rm=T)) %>% 
  mutate(SPAI.group = ifelse(SPAI < median(SPAI, na.rm=T), "low", "high")) %>%
  #mutate(SPAI.group = SPAI %>% ntile(3)) %>% 
  summarise(.by=SPAI.group, cor = cor.test(lds, ms.diag) %>% apa::cor_apa(r_ci=T, print=F)) %>% arrange(SPAI.group)
#for low social anxiety: faster diagnostic dwell <=> MORE fear generalization (LESS heart rate DECELERATION)


#SPAI x ms.diag.type
print(data.wide %>% filter(subject %in% exclusions.eye.ms == F) %>% #manual exclusion because of extreme latency
        gather("lds.type", "lds", c("HR_Gen_eyes_lds", "HR_Gen_mn_lds")) %>%
        gather("ms.diag.type", "ms.diag", c("Gen_eyes_ms", "Gen_mn_ms")) %>%
        select(subject:STAI, lds.type:ms.diag) %>% 
        mutate(lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
               ms.diag.type = ifelse(grepl("_eyes_", ms.diag.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>%
        filter(lds.type == ms.diag.type) %>%
        ggplot(aes(x=SPAI, y=lds, color=ms.diag.type, fill=ms.diag.type)) +
        #geom_line(aes(group=subject), color="black", linetype="dashed") +
        geom_smooth(method="lm", size=1.5, alpha = .2) +
        geom_point(size=2) +
        ylab("LDS of Heart Change") + labs(color="Diagnostic", fill="Diagnostic") +
        myGgTheme +
        theme(
          #legend.position = "none",
          #legend.position = c(.87, .87),
          legend.position = c(1-.87, 1-.87),
        )
)
data.wide %>% with(cor.test(HR_Gen_eyes_lds, SPAI, alternative="two.sided")) %>% correlation_out()
data.wide %>% with(cor.test(HR_Gen_mn_lds, SPAI, alternative="two.sided")) %>% correlation_out()
#correlation SPAI & heart LDS: high generalization <=> high anxiety for diagnostic eyes
#                              low  generalization <=> high anxiety for diagnostic mouth/nose
#interpretation complicated by biphasic heart rate pattern of highly anxious participants


# EDA (no effects) --------------------------------------
reg.eda.dwell = data.wide %>% gather("lds.type", "lds", c("EDA_Gen_eyes_lds", "EDA_Gen_mn_lds")) %>% 
  gather("dwell.type", "dwell", c("Gen_eyes_dwell", "Gen_mn_dwell")) %>%
  select(subject:STAI, lds.type:dwell) %>% 
  mutate(lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
         dwell.type = ifelse(grepl("_eyes_", dwell.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
  filter(lds.type == dwell.type) %>% 
  #mutate(dwell.type = factor(dwell.type)) %>% 
  mutate(dwell.type = ifelse(dwell.type=="Eyes", -1, 1)) %>% 
  mutate(lds = scale(lds), dwell = scale(dwell), SPAI = scale(SPAI), STAI = scale(STAI)) %>% #z-transform
  lmer(lds ~ dwell*dwell.type*SPAI + (1|subject), .) 
  #lmer(lds ~ dwell*dwell.type*STAI + (1|subject), .) 
  #lmer(lds ~ dwell*dwell.type + (1|subject), .) 

reg.eda.dwell %>% summary() %>% print()
reg.eda.dwell %>% lmer.ci()
#reg.eda.dwell %>% lmer.ci(twotailed = F)

data.wide %>% gather("lds.type", "lds", c("EDA_Gen_eyes_lds", "EDA_Gen_mn_lds")) %>% 
  gather("dwell.type", "dwell", c("Gen_eyes_dwell", "Gen_mn_dwell")) %>%
  select(lds.type:dwell) %>% 
  mutate(lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
         dwell.type = ifelse(grepl("_eyes_", dwell.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
  filter(lds.type == dwell.type) %>% 
  ggplot(aes(x=dwell, y=lds, color=dwell.type, fill=dwell.type)) +
  geom_smooth(method="lm", size=1.5, alpha = .2) +
  geom_point(size=2) +
  ylab("Linear Deviation Score (EDA)") + xlab("Diagnostic Dwell (%)") + labs(color="Diagnostic", fill="Diagnostic") +
  myGgTheme + theme(
    #legend.position = "none",
    legend.position = c(1-.87, .87)
  )


#EDA: LDS & Time to diagnostic ROI
reg.eda.ms = data.wide %>% gather("lds.type", "lds", c("EDA_Gen_eyes_lds", "EDA_Gen_mn_lds")) %>% 
  gather("ms.diag.type", "ms.diag", c("Gen_eyes_ms", "Gen_mn_ms")) %>%
  select(subject:STAI, lds.type:ms.diag) %>% 
  mutate(ms.diag = sqrt(ms.diag),
         lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
         ms.diag.type = ifelse(grepl("_eyes_", ms.diag.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
  na.omit() %>% 
  filter(lds.type == ms.diag.type) %>% 
  mutate(ms.diag.type = ifelse(ms.diag.type=="Eyes", -1, 1)) %>% 
  mutate(lds = scale(lds), ms.diag = scale(ms.diag), SPAI = scale(SPAI), STAI = scale(STAI)) %>% #z-transform
  lmer(lds ~ ms.diag*ms.diag.type*SPAI + (1|subject), .) 
  #lmer(lds ~ ms.diag*ms.diag.type*STAI + (1|subject), .) 
  #lmer(lds ~ ms.diag*ms.diag.type + (1|subject), .) 

reg.eda.ms %>% summary() %>% print()
reg.eda.ms %>% lmer.ci()
reg.eda.ms %>% lmer.ci(twotailed = F)

data.wide %>% gather("lds.type", "lds", c("EDA_Gen_eyes_lds", "EDA_Gen_mn_lds")) %>% 
  gather("ms.diag.type", "ms.diag", c("Gen_eyes_ms", "Gen_mn_ms")) %>%
  select(lds.type:ms.diag) %>% 
  mutate(lds.type = ifelse(grepl("_eyes_", lds.type), "Eyes", "Mouth/Nose") %>% as.factor(),
         ms.diag.type = ifelse(grepl("_eyes_", ms.diag.type), "Eyes", "Mouth/Nose") %>% as.factor()) %>% 
  filter(lds.type == ms.diag.type) %>% 
  ggplot(aes(x=sqrt(ms.diag), y=lds, color=ms.diag.type, fill=ms.diag.type)) +
  geom_smooth(method="lm", size=1.5, alpha = .2) +
  #stat_smooth(method="lm", size=1.5, alpha = .2, fullrange = T) +
  geom_point(size=2) +
  ylab("Linear Deviation Score (EDA)") + xlab(expression("Time to Diagnostic ROI (" * sqrt(ms) * ")")) + labs(color="Diagnostic", fill="Diagnostic") +
  myGgTheme + theme(
    #legend.position = "none",
    legend.position = c(.87, .87))
